requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
flask
flask-cors
mysql-connector-python
python-dotenv